Change Core Data objects and fix PillStats bug (#64)
* Revertes Core Data objects to default scalar versions
* Fixes minor PillStats bug
* Updates Notifications Unit test; 
* Updates jazzy command on README.md
* Fixes notification manager showing outdated notifications

+++ Malaria Prevention App/ViewControllers/Info Hub and Games/User Profile/UserProfilePillTableViewCell.swift
@@ -31,7 +31,7 @@ class UserProfilePillTableViewCell: UITableViewCell {
   
   func updateCell(delegate: SaveRemainingPillsProtocol,
                                 name: String,
-                                remainingMedicine: Int64,
+                                remainingMedicine: Int,
                                 indexPath: NSIndexPath) {
     pillName.text = name
     quantityTextField.text = String(remainingMedicine)
